fix(multi-predictor-synth/E6-S4): forbid float fusion in synth capture
TestProfile_PreModelsCaptureIsByteIdentical has failed on linux/amd64 on
every commit of this branch since bf2f778 (14 commits) while passing on
darwin/arm64. The cause is FMA contraction: Go permits `a + b*c` to be
fused into a single multiply-add, the arm64 backend does it and the amd64
backend does not, so the same source rounds differently per CPU.

That made `pulse profile create`'s byte-identity promise hold per-MACHINE
only, which is worthless for an artefact users commit, diff and share.
E2-S4 closed the map-ordering dimension of the same claim; this is the
second dimension.

Two independent causes, both closed:

- The test's own cohort generator. modelsCohort built its spend/visits
  columns as a chain of fusible float terms, so the INPUT COHORT BYTES
  differed between architectures — 115 of 1,200 rows. It now accumulates
  in exact integer hundredths and divides once, which cannot be undone by
  a later edit that merely looks tidier.
- The capture and generation arithmetic. Every product feeding an add or
  subtract now carries an explicit float64() conversion, the only
  construct the spec defines as forbidding contraction. The compiler
  confirms zero remaining contraction sites in synth/ non-test code
  (-d=fmahash=vy). math.FMA was rejected: it forces fusion everywhere,
  changing amd64's values rather than preserving them.

The shared variance form now lives once, as synth/moments.go
sampleVariance, which also documents the whole rule. The algorithm is
unchanged — the catastrophic-cancellation hazard in (sumSq - mean*sum) is
a separate concern and is not touched here.

synth/moments_internal_test.go is the build-failing gate: sampleVariance,
pearson and computePercentiles each compared against a //go:noinline twin
holding the pre-barrier source, asserting production matches the
BARRIERED value, plus an anti-vacuity check that the twins actually
disagree. Verified to fail on arm64 with any of the three barriers
removed. It can only detect on a contracting architecture, so it logs
which case it is in rather than implying coverage on amd64.

testdata/profile_pre_models.json is regenerated (now via `-update` on the
test, so its provenance is a runnable command) and its comment rewritten
— it is no longer the pre-change tree's output verbatim and must not
claim to be. The diff is 27 changed lines, all float-valued, max
|relative delta| 8.2e-15, no key added or removed, no structural change,
warnings identical. preModelsSpecHash moves as an unavoidable consequence
(a content hash over floats cannot be invariant to a last-bit change);
the constant's comment records the old value and why this one movement is
not a semantic change.

Known residuals, stated in the docs rather than papered over: math.Exp /
math.Log are architecture-specific in the standard library, so a
--fit-shape `shape` section can still differ in its last bits, and
processing/regression (the OLS engine behind `models`) has not been made
fusion-free.

+++ synth/conditional_sample.go
@@ -211,7 +211,7 @@ func (c *categoricalNumericPairSampler) transform(rng *rand.Rand, row map[string
 		}
 		mom = c.fallback
 	}
-	v := mom.mean + rng.NormFloat64()*mom.std
+	v := mom.mean + float64(rng.NormFloat64()*mom.std)
 	if c.hasClamp {
 		if v < c.clampMin {
 			v = c.clampMin
@@ -402,7 +402,7 @@ func (s *setNumericPairSampler) transform(rng *rand.Rand, row map[string]any) {
 		}
 		mom = s.fallback
 	}
-	v := mom.mean + rng.NormFloat64()*mom.std
+	v := mom.mean + float64(rng.NormFloat64()*mom.std)
 	if s.hasClamp {
 		if v < s.clampMin {
 			v = s.clampMin
